functions/alert_notifier/requirements.txt
functions-framework==3.5.0
google-cloud-firestore==2.18.0
